wallet/requirements.txt
Django==4.1.1
django_filter==22.1
djangorestframework==3.14.0
drf_excel==2.2.0
psycopg2==2.9.3
djoser==2.1.0